Allow pool creation for radosgw from Ansible
Radosgw is able to create the needed pools on first use. But this is
quite limited as you can't set the number of placement groups or CRUSH
rules to use. In addition the radosgw needs to have permissions to
manipulate pools for this to work. Which is not desirable for production
use.

This allows pool creation to be done by the Ansible playbook on the
first monitor host. This setting is off by default.